refactor endpoints out to individual routers
